backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==0.19.1
duckduckgo-search==2.9.5
openai==1.3.0
python-multipart==0.0.6
